+++ docs/changelog.rst
@@ -10,12 +10,13 @@ a. It is now possible to disable one or two out of the three thresholds in the :
 b. The :py:func:`yasa.spindles_detect` now returns the timing (in seconds) of the most prominent peak of each spindles (``Peak``). This can be used to perform spindles-SO coupling, as explained in the Jupyter notebooks.
 c. The yasa.get_sync_sw has been renamed to :py:func:`yasa.get_sync_events` and several arguments have been renamed to accomodate the output from the spindles detection as well as the slow-waves detection (for which the function was originally intended).
 d. Added a notebook on spindles-SO coupling.
+e. Removed Travis and AppVeyor testing for Python 3.5.
 
 v0.1.8 (October 2019)
 ---------------------
 
 a. Added :py:func:`yasa.plot_spectrogram` function.
-b. Added `lspopt <https://github.com/hbldh/lspopt>`_ in the depencencies
+b. Added `lspopt <https://github.com/hbldh/lspopt>`_ in the dependencies.
 c. YASA now requires `MNE <https://mne.tools/stable/index.html>`_>0.19.
 d. Added a notebook on non-linear features.